example_extension_project/requirements.txt
django-coverage-plugin==1.5.0
django-extensions==1.9.1
factory-boy==2.9.2
